+++ src/pages/claim_details/ClaimDetails.css
