+++ src/pages/claim_details/ClaimDetails.tsx
